added fixes on data consistency

+++ spring-08-homework/src/main/java/ru/otus/spring08homework/shell/BookServiceCommands.java
@@ -160,6 +160,9 @@ public class BookServiceCommands {
         if (!authorService.existsById(id)) {
             return "автор не найден";
         }
+        if (!bookService.findByAuthorsContaining(id).isEmpty()) {
+            return ""
+        }
         authorService.deleteById(id);
         return "автор удален";
     }
